# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/center": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user center information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户中心信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功，返回用户信息和用户详情",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "User login with username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功返回token",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "非法请求",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "密码错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user/login-phone": {
            "post": {
                "description": "通过发送的验证码登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "手机号登录",
                "parameters": [
                    {
                        "description": "通过验证码登录",
                        "name": "验证码",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.VerificationCode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功返回token",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "非法请求",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在或者验证码错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user/profile/{uid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user profile by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户资料",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回用户资料",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "非法请求",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update user profile by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更新用户资料",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User profile data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateUser"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated user profile data",
                        "schema": {
                            "$ref": "#/definitions/user.UpdateUser"
                        }
                    },
                    "400": {
                        "description": "Invalid request or malformed JSON",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/user/verification-code/{kind}": {
            "post": {
                "description": "Get verification code for a specific kind",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获得验证码",
                "parameters": [
                    {
                        "enum": [
                            "login",
                            "register",
                            "reset_password"
                        ],
                        "type": "string",
                        "description": "Kind of verification code",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "手机号",
                        "name": "Phone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CodeSendUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功返回token",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "非法请求",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在或者验证码错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "resp.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "default": 1000
                },
                "msg": {
                    "type": "string"
                },
                "time_now": {
                    "type": "integer"
                }
            }
        },
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "default": 0
                },
                "data": {},
                "time_now": {
                    "type": "integer"
                }
            }
        },
        "user.CodeSendUser": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "user.Login": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "username": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 6
                }
            }
        },
...
        "user.UpdateUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "username": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 6
                }
            }
        },
        "user.VerificationCode": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "kind": {
                    "type": "string",
                    "enum": [
                        "login",
                        "register",
                        "reset_password"
                    ]
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    }
}